Minor anti cheating mechanism
You are now unable to modify the score in console with just 1 command.

You can use
"allowScore=true; updateScore(this.score);"
or
"score = 1000; akamai = 1000;"

but running one command first and then the other does not work. because
it tries to fix it in every update cycle so you have to be able to run
them faster than the update cycle